refactored all
