+++ internal/helper/history/save.go
@@ -3,12 +3,12 @@ package historyhelper
 import (
 	"path/filepath"
 
-	appdata "github.com/fiwon123/crower/internal/data/app"
+	"github.com/fiwon123/crower/internal/app"
 	"github.com/fiwon123/crower/pkg/crowerutils"
 )
 
 // Save history data on cfg file
-func Save(app *appdata.Data) {
+func Save(app *app.Config) {
 	data := app.History.GetLast()
 
 	newDataPath := filepath.Join(app.HistoryFolderPath, data.File)
@@ -18,6 +18,6 @@ func Save(app *appdata.Data) {
 	crowerutils.WriteJson(app.History, app.HistoryFilePath)
 }
 
-func SaveOnlyHistory(app *appdata.Data) {
+func SaveOnlyHistory(app *app.Config) {
 	crowerutils.WriteJson(app.History, app.HistoryFilePath)
 }
