+++ internal/helper/inputs.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 	"unicode"
 
+	"github.com/fiwon123/crower/internal/app"
 	"github.com/fiwon123/crower/internal/crowererrors"
-	appdata "github.com/fiwon123/crower/internal/data/app"
 )
 
 const (
@@ -45,7 +45,7 @@ func CheckYesAnswer(input string) bool {
 	return false
 }
 
-func GetUserInput(ask string, fnValid func(string, *appdata.Data) (any, error), app *appdata.Data) any {
+func GetUserInput(ask string, fnValid func(string, *app.Config) (any, error), app *app.Config) any {
 	ok := false
 	input := ""
 	var output any
@@ -68,7 +68,7 @@ func GetUserInput(ask string, fnValid func(string, *appdata.Data) (any, error), 
 	return output
 }
 
-func GetUserConfirmation(ask string, app *appdata.Data) bool {
+func GetUserConfirmation(ask string, app *app.Config) bool {
 	ok := false
 	input := ""
 	var confirmation bool
@@ -104,7 +104,7 @@ func IsValidConfirmation(input string) (bool, error) {
 	return true, nil
 }
 
-func IsValidInput(input string, app *appdata.Data) (any, error) {
+func IsValidInput(input string, app *app.Config) (any, error) {
 	if input == "" {
 		return "", crowererrors.EmptyInput()
 	}
@@ -112,7 +112,7 @@ func IsValidInput(input string, app *appdata.Data) (any, error) {
 	return input, nil
 }
 
-func InputAlias(app *appdata.Data) []string {
+func InputAlias(app *app.Config) []string {
 	output := []string{}
 
 	alias := "none"
@@ -128,7 +128,7 @@ func InputAlias(app *appdata.Data) []string {
 	return output
 }
 
-func InputName(app *appdata.Data) string {
+func InputName(app *app.Config) string {
 	name := ""
 	for name == "" {
 		name = GetUserInput("Add new name: ", IsValidInput, app).(string)
@@ -137,7 +137,7 @@ func InputName(app *appdata.Data) string {
 	return name
 }
 
-func InputExec(app *appdata.Data) string {
+func InputExec(app *app.Config) string {
 	exec := ""
 	for exec == "" {
 		exec = GetUserInput("Add new exec: ", IsValidInput, app).(string)
@@ -146,7 +146,7 @@ func InputExec(app *appdata.Data) string {
 	return exec
 }
 
-func isValidAlias(input string, app *appdata.Data) (any, error) {
+func isValidAlias(input string, app *app.Config) (any, error) {
 	for _, r := range input {
 		if !unicode.IsNumber(r) && !unicode.IsLetter(r) {
 			return "", crowererrors.OnlyLettersAndNumbers()
@@ -156,7 +156,7 @@ func isValidAlias(input string, app *appdata.Data) (any, error) {
 	return input, nil
 }
 
-func IsValidInputKey(input string, app *appdata.Data) (any, error) {
+func IsValidInputKey(input string, app *app.Config) (any, error) {
 	index, err := strconv.Atoi(input)
 	if err != nil {
 		return "", crowererrors.InvalidRows()
@@ -169,7 +169,7 @@ func IsValidInputKey(input string, app *appdata.Data) (any, error) {
 	return app.OrderKeys[index], nil
 }
 
-func IsValidContentKey(input string, app *appdata.Data) (any, error) {
+func IsValidContentKey(input string, app *app.Config) (any, error) {
 	index, err := strconv.Atoi(input)
 	if err != nil {
 		return "", crowererrors.InvalidRows()
